Add Kernel test base class and update tests
