+++ tests/src/Kernel/FileLinkUsageScannerTest.php
@@ -2,17 +2,16 @@
 
 namespace Drupal\Tests\filelink_usage\Kernel;
 
-use Drupal\KernelTests\KernelTestBase;
+use Drupal\Tests\filelink_usage\Kernel\FileLinkUsageKernelTestBase;
 use Drupal\file\Entity\File;
 use Drupal\node\Entity\Node;
-use Drupal\node\Entity\NodeType;
 
 /**
  * Ensures file usage entries are not duplicated when scanning.
  *
  * @group filelink_usage
  */
-class FileLinkUsageScannerTest extends KernelTestBase {
+class FileLinkUsageScannerTest extends FileLinkUsageKernelTestBase {
 
   /**
    * {@inheritdoc}
@@ -33,24 +32,6 @@ class FileLinkUsageScannerTest extends KernelTestBase {
    */
   protected function setUp(): void {
     parent::setUp();
-
-    $this->installEntitySchema('user');
-    $this->installEntitySchema('file');
-    $this->installEntitySchema('node');
-    $this->installSchema('file', ['file_usage']);
-    $this->installSchema('filelink_usage', [
-      'filelink_usage_matches',
-      'filelink_usage_scan_status',
-    ]);
-    $this->installSchema('node', ['node_access']);
-    $this->installConfig(['system', 'node', 'filter']);
-
-    $node_type = NodeType::create([
-      'type' => 'article',
-      'name' => 'Article',
-    ]);
-    $node_type->save();
-    node_add_body_field($node_type);
   }
 
   /**
